requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysql-connector-python==8.0.31
protobuf==3.20.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
xlwt==1.3.0
